Create README.md
With the commission structure being re-vamped by a certain red wireless company, this app will make it easy to quickly see how much of a payout a rep can expect from a single transaction.  The app will also let you track a running total of sales for a particular month, giving a rep the ability to compare their manually inputted totals to the employer-provided commission tracker.